mooore clean code

+++ src/main/java/com/nonogram_online/buttons/BasicButton.java
@@ -136,7 +136,7 @@ public class BasicButton extends JPanel {
     }
 
     protected void setFontSize() {
-        fontSize = (int) (width / 12);
+        fontSize = (width / 12);
     }
 
     @Override
